Separates nbe instance and syntax from main file
