+++ src/main/scala/core/Statelesser.scala
@@ -1,10 +1,6 @@
 package statelesser
 package core
 
-import Function.const
-import scalaz._, Scalaz._
-import Leibniz._
-
 import optic._
 
 trait Statelesser[Expr[_]] {
@@ -74,242 +70,8 @@ object Statelesser {
 
   implicit def metacircular = new interpreter.Metacircular
 
-  implicit def tsemantic: Statelesser[Semantic] = new Statelesser[Semantic] {
-
-    def gtVert[S, A, B](
-        usem: Semantic[Getter[S, A]],
-        dsem: Semantic[Getter[A, B]]) =
-      vert(usem, dsem)
+  implicit def tsemantic = new interpreter.NBE
  
-    def gtHori[S, A, B](
-        lsem: Semantic[Getter[S, A]],
-        rsem: Semantic[Getter[S, B]]): Semantic[Getter[S, (A, B)]] =
-      cart(lsem, rsem)
-
-    def aflVert[S, A, B](
-        usem: Semantic[AffineFold[S, A]],
-        dsem: Semantic[AffineFold[A, B]]) =
-      vert(usem, dsem)
- 
-    def aflHori[S, A, B](
-        lsem: Semantic[AffineFold[S, A]],
-        rsem: Semantic[AffineFold[S, B]]) =
-      cart(lsem, rsem)
- 
-    def flVert[S, A, B](
-        usem: Semantic[Fold[S, A]],
-        dsem: Semantic[Fold[A, B]]) =
-      vert(usem, dsem)
- 
-    def flHori[S, A, B](
-        lsem: Semantic[Fold[S, A]],
-        rsem: Semantic[Fold[S, B]]): Semantic[Fold[S, (A, B)]] =
-      cart(lsem, rsem)
-
-    def filtered[S](p: Semantic[Getter[S, Boolean]]) =
-      p.map(sem => new Todo[AffineFold, S, S] {
-        def apply[T](done: Done[AffineFold, T, S]) = sem match {
-          case todo: Todo[Getter, S, Boolean] => {
-            // We're just selecting a unique Boolean, so must be `Just` 
-            // XXX: weird, turning an `AffineFold` into a `Getter`. It's just a
-            // phantom type so it won't fail, but conceptually, this is wrong.
-            val Just(e) = todo(Done[Getter, T, S](
-              done.expr, 
-              done.filt, 
-              done.vars)).expr
-            done.copy(filt = done.filt + e)
-          }
-        }
-      })
-
-    def sub: Semantic[Getter[(Int, Int), Int]] =
-      state(new Todo[Getter, (Int, Int), Int] {
-        def apply[S](done: Done[Getter, S, (Int, Int)]) = done.expr match {
-          case pair: Pair[S, (Int, Int), Int, Int]@unchecked =>
-            Done(Just((pair.l, pair.r) match {
-              case (Just(l), Just(LikeInt(0))) => l
-              case (Just(LikeInt(x)), Just(LikeInt(y))) => LikeInt(x - y)
-              case (Just(l), Just(r)) => Sub(l, r, implicitly)
-            }), done.filt, done.vars)
-        }
-      })
-
-    def greaterThan: Semantic[Getter[(Int, Int), Boolean]] =
-      state(new Todo[Getter, (Int, Int), Boolean] {
-        def apply[S](done: Done[Getter, S, (Int, Int)]) = done.expr match {
-          case pair: Pair[S, (Int, Int), Int, Int]@unchecked =>
-            Done(Just((pair.l, pair.r) match {
-              case (Just(LikeInt(x)), Just(LikeInt(y))) => LikeBool(x > y)
-              case (Just(l), Just(r)) => Gt(l, r, implicitly)
-            }), done.filt, done.vars)
-        }
-      })
-
-    def first[A, B]: Semantic[Getter[(A, B), A]] =
-      state(new Todo[Getter, (A, B), A] {
-        def apply[S](done: Done[Getter, S, (A, B)]) = done.expr match {
-          case pair: Pair[S, (A, B), A, B]@unchecked =>
-            cleanUnusedVars(Done(pair.l, done.filt, done.vars))
-        }
-      })
-
-    def second[A, B]: Semantic[Getter[(A, B), B]] =
-      state(new Todo[Getter, (A, B), B] {
-        def apply[S](done: Done[Getter, S, (A, B)]) = done.expr match {
-          case pair: Pair[S, (A, B), A, B]@unchecked =>
-            cleanUnusedVars(Done(pair.r, done.filt, done.vars))
-        }
-      })
-
-    def likeInt[S](i: Int): Semantic[Getter[S, Int]] =
-      state(new Todo[Getter, S, Int] {
-        def apply[T](done: Done[Getter, T, S]) =
-          Done(Just(LikeInt(i)), done.filt, done.vars)
-      })
- 
-    def likeBool[S](b: Boolean): Semantic[Getter[S, Boolean]] =
-      state(new Todo[Getter, S, Boolean] {
-        def apply[T](done: Done[Getter, T, S]) =
-          Done(Just(LikeBool(b)), done.filt, done.vars)
-      })
-
-    def likeStr[S](s: String): Semantic[Getter[S, String]] =
-      state(new Todo[Getter, S, String] {
-        def apply[T](done: Done[Getter, T, S]) =
-          Done(Just(LikeStr(s)), done.filt, done.vars)
-      })
-
-    def id[S]: Semantic[Getter[S, S]] =
-      state(new Todo[Getter, S, S] {
-        def apply[T](done: Done[Getter, T, S]) = done
-      })
-
-    def not: Semantic[Getter[Boolean, Boolean]] =
-      state(new Todo[Getter, Boolean, Boolean] {
-        def apply[S](done: Done[Getter, S, Boolean]) = done.expr match {
-          case just: Just[S, Boolean] =>
-            Done(Just(just.e match {
-              case LikeBool(b) => LikeBool(!b)
-              case Not(b, _) => b
-            }), done.filt, done.vars)
-        }
-      })
-
-    def gtAsAfl[S, A](gt: Semantic[Getter[S, A]]) = 
-      gt.map(_ match {
-        case done: Done[Getter, S, A] => done.as[AffineFold]
-        case todo: Todo[Getter, S, A] => todo.as[AffineFold]
-      })
-
-    def aflAsFl[S, A](afl: Semantic[AffineFold[S, A]]) = 
-      afl.map(_ match {
-        case done: Done[AffineFold, S, A] => done.as[Fold]
-        case todo: Todo[AffineFold, S, A] => todo.as[Fold]
-      })
-
-    private def vert[O[_, _], S, A, B](
-        usem: Semantic[O[S, A]],
-        dsem: Semantic[O[A, B]]): Semantic[O[S, B]] = 
-      for {
-        sem1 <- usem
-        sem2 <- dsem
-      } yield (sem1, sem2) match {
-        case (x, todo: Todo[O, A, B]@unchecked) => todo compose x
-        case _ => throw new Error(s"Can't vert compose with 'Done' semantic: $sem2")
-      }
-
-    private def unifyVars[O[_, _], S, A](done: Done[O, S, A]): Done[O, S, A] = {
-      val (t, rws) = done.vars
-        .groupBy(_._2)
-        .map { case (k, v) => (k, v.keys.toList) }
-        .foldLeft((Map.empty[Symbol, statelesser.Value], Set.empty[(String, String)])) {
-          case ((m, rws), (v, k :: ks)) =>
-            (m + (k -> v), rws ++ ks.map((_, k)))
-        }
-      if (rws.isEmpty) done
-      else unifyVars(Done(
-        done.expr.renameVars(rws), 
-        done.filt.map(_.renameVars(rws)), 
-        t))
-    }
-
-    private def cleanUnusedVars[O[_, _], S, A](
-        done: Done[O, S, A]): Done[O, S, A] = {
-      def deps(k: String): Set[String] =
-        done.vars(k).fold(const(Set()), sel => deps(sel.v.sym) + sel.v.sym)
-      val used: Set[String] =
-        done.expr.vars ++ done.filt.flatMap(_.vars)
-      done.copy(vars = done.vars
-        .filterKeys((used ++ (used.flatMap(deps))).contains(_)))
-    }
-
-    private def cart[O[_, _], S, A, B](
-        lsem: Semantic[O[S, A]],
-        rsem: Semantic[O[S, B]]): Semantic[O[S, (A, B)]] =
-      for {
-        sem1 <- lsem
-        sem2 <- rsem
-      } yield (sem1, sem2) match {
-        case (ltodo: Todo[O, S, A], rtodo: Todo[O, S, B]) => 
-          new Todo[O, S, (A, B)] {
-            def apply[T](done: Done[O, T, S]) = (ltodo(done), rtodo(done)) match {
-              case (Done(le, lf, lv), Done(re, rf, rv)) =>
-                unifyVars(Done[O, T, (A, B)](
-                  Pair[T, (A, B), A, B](le, re, implicitly), 
-                  lf ++ rf, 
-                  lv ++ rv))
-            }
-          }
-        case (Done(le, lf, lv), Done(re, rf, rv)) =>
-          Done(Pair[S, (A, B), A, B](le, re, implicitly), lf ++ rf, lv ++ rv)
-      }
-  }
- 
-  def fresh: State[Stream[String], String] =
-    for {
-      s <- get[Stream[String]]
-      _ <- modify[Stream[String]](_.tail)
-    } yield s.head
-
-  trait Syntax {
-
-    implicit class FoldOps[Expr[_], S, A](
-        l: Expr[Fold[S, A]])(implicit
-        O: Statelesser[Expr]) {
-      def >[B](r: Expr[Fold[A, B]]): Expr[Fold[S, B]] = O.flVert(l, r)
-      def *[B](r: Expr[Fold[S, B]]): Expr[Fold[S, (A, B)]] = O.flHori(l, r)
-    }
-
-    implicit class AffineFoldOps[Expr[_], S, A](
-        l: Expr[AffineFold[S, A]])(implicit
-        O: Statelesser[Expr]) {
-      def asFold: Expr[Fold[S, A]] = O.aflAsFl(l)
-      def >[B](r: Expr[AffineFold[A, B]]): Expr[AffineFold[S, B]] =
-        O.aflVert(l, r)
-      def *[B](r: Expr[AffineFold[S, B]]): Expr[AffineFold[S, (A, B)]] =
-        O.aflHori(l, r)
-      def <*[B](r: Expr[AffineFold[S, B]]): Expr[AffineFold[S, A]] =
-        l * r > O.first.asAffineFold
-      def *>[B](r: Expr[AffineFold[S, B]]): Expr[AffineFold[S, B]] =
-        l * r > O.second.asAffineFold
-    }
-
-    implicit class GetterOps[Expr[_], S, A](
-        l: Expr[Getter[S, A]])(implicit
-        O: Statelesser[Expr]) {
-      def asAffineFold: Expr[AffineFold[S, A]] = O.gtAsAfl(l)
-      def asFold: Expr[Fold[S, A]] = l.asAffineFold.asFold
-      def >[B](r: Expr[Getter[A, B]]): Expr[Getter[S, B]] = O.gtVert(l, r)
-      def *[B](r: Expr[Getter[S, B]]): Expr[Getter[S, (A, B)]] = O.gtHori(l, r)
-    }
-
-    implicit class IntGetterOps[Expr[_], S](
-        l: Expr[Getter[S, Int]])(implicit
-        O: Statelesser[Expr]) {
-      def -(r: Expr[Getter[S, Int]]): Expr[Getter[S, Int]] = l * r > O.sub
-    }
-  }
-
   object syntax extends Syntax
 }
